homework week 5
In class, we added code to analysis.Rmd to create summary statistics for the participant sample, i.e., mean age and SD age, and to print this as a table. If yours isn't working, a working example of this can be found in my github repo. Modify your own processing.Rmd and analysis.Rmd code as necessary to get it working.

Add to analysis.Rmd in order to create:

- A table of the overall sample size.

- A table with both counts and percentages of each gender in the sample.

- A table with the mean and standard deviation of mean_evaluation, a) for the overall sample, and b) for each gender category.

Add to processing.Rmd in order to:

- Create a new data frame under the "Exclusions / data quality" heading. This data frame should take the data_amp_raw object, exclude instructions and practice trials, create a new variable which indicates whether the latency on each trial was problematically low (< 100ms) or not,